+++ scripts/macros/2024/classFeatures/pugilist/squaredCircle/heavyweight.js
@@ -1,0 +1,5 @@
+export let heavyweight = {
+    name: 'Heavyweight',
+    version: '1.5.32',
+    rules: 'modern'
+};